Theme and tree-sitter syntax highlighting support (#23)
* Theme WIP

* More progress on theme parsing

* Full theme and highlighting support (no LSP support yet)

* fixed incorrectly colored status line separators

* Added more themes

* Removed unused vars

* Some clippy linting (some of which is temporary

* Fixed persistent mouse capture issue

* Updated README installation steps

* Fixed unit test fixture and other clippy warnings

* Removed tests that are not necessary at this time

+++ src/buffer.rs
@@ -48,9 +48,9 @@ impl Buffer {
         self.lines.len()
     }
 
-    pub fn insert(&mut self, x: u16, y: usize, c: char) {
+    pub fn insert(&mut self, x: usize, y: usize, c: char) {
         if let Some(line) = self.lines.get_mut(y) {
-            (*line).insert(x as usize, c);
+            (*line).insert(x, c);
         }
     }
 
@@ -58,9 +58,9 @@ impl Buffer {
         self.lines.insert(line, content);
     }
 
-    pub fn remove(&mut self, x: u16, y: usize) {
+    pub fn remove(&mut self, x: usize, y: usize) {
         if let Some(line) = self.lines.get_mut(y) {
-            (*line).remove(x as usize);
+            (*line).remove(x);
         }
     }
